refactor(frontend): align mobile settings ownership
