+++ apps/mobile/src/hooks/access/cloud/repositories/use-mobile-repositories.ts
@@ -1,0 +1,12 @@
+import {
+  useCloudGitRepositories,
+  useSaveCloudRepoConfig,
+} from "@proliferate/cloud-sdk-react";
+
+export function useMobileGitRepositories(visible: boolean) {
+  return useCloudGitRepositories({}, visible);
+}
+
+export function useSaveMobileRepoConfig() {
+  return useSaveCloudRepoConfig();
+}
