+++ apps/mobile/src/lib/access/native/open-url.ts
@@ -1,0 +1,5 @@
+import { Linking } from "react-native";
+
+export function openNativeUrl(url: string): Promise<void> {
+  return Linking.openURL(url);
+}
